lib/stroll/requirements.txt
paho-mqtt==1.4.0
Adafruit-DHT==1.4.0
